Refactor orthogonal_label_generation.py to use label_generation_utils.py and improve performance/stability
